+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - XOT-1280 - update components for cookie banner
 
 ### Implemented enhancements
+- CI-631 - Update business environment guide form and content in form consent sections
 - CI-653 - Add opportunities search link to the header
 - CI-651 - Alter text alignment and add vignettes to hero components
 - XOT-1262 - Update to new field name on opportunity page. Remove logic that is now in the cms
@@ -12,7 +13,6 @@
 - CI-649 - add missing field to invest contact form
 - CI-644 - Update capital invest form
 - CI-649 - Update invest contact form fields, email options and tests
-- CI-631 - Update business environment guide form
 - CI-645 - Add translations for 'About us'
 
 ### Bugs fixed